+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- **[iOS]** `QuillEditor` doesn't respect the system keyboard brightness by default.
+
 ## [11.1.1] - 2025-03-19
 
 ### Fixed